requirements.txt
Pillow
pillow-heif
opencv-python
imagehash